Move ShardingItemParameters (#1195)
